+++ packages/ui/Scrollable/ScrollableNative.js
@@ -253,6 +253,14 @@ class ScrollableBaseNative extends Component {
 		onScrollStop: PropTypes.func,
 
 		/**
+		 * Called when invalidating [ScrollerNative]{@link ui/Scroller.ScrollerNative}'s bounds
+		 *
+		 * @type {function}
+		 * @private
+		 */
+		onUpdate: PropTypes.func,
+
+		/**
 		 * Called when wheeling.
 		 *
 		 * @type {Function}
@@ -328,6 +336,7 @@ class ScrollableBaseNative extends Component {
 		onScroll: nop,
 		onScrollStart: nop,
 		onScrollStop: nop,
+		onUpdate: nop,
 		verticalScrollbar: 'auto'
 	}
 
@@ -386,6 +395,7 @@ class ScrollableBaseNative extends Component {
 
 	componentDidUpdate (prevProps, prevState) {
 		const
+			{onUpdate} = this.props,
 			{isHorizontalScrollbarVisible, isVerticalScrollbarVisible} = this.state,
 			{hasDataSizeChanged} = this.childRef;
 
@@ -403,10 +413,15 @@ class ScrollableBaseNative extends Component {
 			hasDataSizeChanged === false &&
 			(isHorizontalScrollbarVisible && !prevState.isHorizontalScrollbarVisible || isVerticalScrollbarVisible && !prevState.isVerticalScrollbarVisible)
 		) {
+			// For VirtualList
 			this.deferScrollTo = false;
 			this.isUpdatedScrollThumb = this.updateScrollThumbSize();
-		} else {
-			this.updateScrollbars();
+		} else if (!this.updateScrollbars()) {
+			if (this.shouldUpdateScroller && !this.childRef.hasOwnProperty('hasDataSizeChanged')) {
+				// For Scroller
+				onUpdate();
+			}
+			this.shouldUpdateScroller = false;
 		}
 
 		if (this.scrollToInfo !== null) {
@@ -445,6 +460,7 @@ class ScrollableBaseNative extends Component {
 	// TODO: consider replacing forceUpdate() by storing bounds in state rather than a non-
 	// state member.
 	enqueueForceUpdate = () => {
+		this.shouldUpdateScroller = true;
 		this.childRef.calculateMetrics();
 		this.forceUpdate();
 	}
@@ -468,6 +484,7 @@ class ScrollableBaseNative extends Component {
 	deferScrollTo = true
 	isScrollAnimationTargetAccumulated = false
 	isUpdatedScrollThumb = false
+	shouldUpdateScroller = false
 
 	// overscroll
 	overscrollStatus = {
@@ -1138,6 +1155,8 @@ class ScrollableBaseNative extends Component {
 			this.deferScrollTo = false;
 			this.isUpdatedScrollThumb = this.updateScrollThumbSize();
 		}
+
+		return isVisibilityChanged;
 	}
 
 	updateScrollThumbSize = () => {
@@ -1266,6 +1285,7 @@ class ScrollableBaseNative extends Component {
 		delete rest.onScroll;
 		delete rest.onScrollStart;
 		delete rest.onScrollStop;
+		delete rest.onUpdate;
 		delete rest.onWheel;
 		delete rest.removeEventListeners;
 		delete rest.scrollStopOnScroll;